test: split declaration-order cases by purpose
