+++ tests/all.test.js
@@ -1,0 +1,3 @@
+import "./declaration-order/core.test.js";
+import "./declaration-order/options.test.js";
+import "./declaration-order/regressions.test.js";
